Implement saving and deleting of movies watched
Room database is now in the early stages of being implemented - if a
movie is checked then the room db is checked to see if film exists in
it already. If it does, film is removed from db, otherwise it is added
to db indicating it's been watched.

Also, when loading film-list recyclerview room db is queried to see if
films exist in the db indicating they've been watched. If they have,
background turns green and check mark is brightened, otherwise it goes
to default movie bg. (Room DB implementation is still a WIP)